requirements.txt
streamlit
pandas
matplotlib
seaborn
scikit-learn
umap-learn
transformers
torch
tqdm
openai